requirements.txt
# Medical Center AI Chatbot - Complete Requirements
# Install with: pip install -r requirements.txt

# Core Web Framework
flask==3.0.0
flask-cors==4.0.0

# Environment Configuration
python-dotenv==1.2.1

# CrewAI Framework
crewai==1.4.1
crewai-tools==1.4.1

# LiteLLM for OpenRouter API integration
litellm>=1.40.0

# Vector Database
qdrant-client==1.11.3

# Data Processing
pandas==2.2.0
openpyxl==3.1.5
PyPDF2==3.0.1

# CrewAI Tools Dependencies
beautifulsoup4>=4.13.4
docker>=7.1.0
lancedb>=0.5.4
pypdf>=5.9.0
python-docx>=1.2.0
pytube>=15.0.0
tiktoken>=0.8.0
youtube-transcript-api>=1.2.2

# HTTP Requests
requests>=2.32.3

# Data Validation
pydantic>=2.12.0

# LangChain for embeddings (using Ollama API)
langchain==0.3.0 
langchain-community==0.3.0 
langchain-text-splitters==0.3.0
langchain-core==1.0.4
langchain-ollama==1.0.0  # For local embeddings